fix: fix auth channels bugs from merged PR
