+++ frontend/.prettierignore
@@ -33,3 +33,4 @@ package-lock.json
 pnpm-lock.yaml
 
 .storybook
+routeTree.gen.ts
