+++ frontend/src/routes/otp/_layout.tsx
@@ -1,7 +1,7 @@
 import { createFileRoute, Outlet } from '@tanstack/react-router'
 
 import { OTPSidebar } from '@/components/sidebar/otp-sidebar'
-import { SidebarInput } from '@/components/ui/sidebar'
+import { SidebarInset } from '@/components/ui/sidebar'
 import { SidebarProvider } from '@/providers/sidebar-provider'
 
 export const Route = createFileRoute('/otp/_layout')({
@@ -13,11 +13,11 @@ function RouteComponent() {
     <div className="wrap flex w-full min-h-screen relative">
       <SidebarProvider>
         <OTPSidebar />
-        <SidebarInput>
+        <SidebarInset>
           <main className="p-6 flex-1">
             <Outlet />
           </main>
-        </SidebarInput>
+        </SidebarInset>
       </SidebarProvider>
     </div>
   )
